Remove WaitForQueryDriver
