+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/query/SQSQueryDriver.java
@@ -16,28 +16,63 @@
 
 package sleeper.systemtest.drivers.query;
 
+import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.sqs.AmazonSQS;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import sleeper.core.util.PollWithRetries;
 import sleeper.query.model.Query;
 import sleeper.query.model.QuerySerDe;
+import sleeper.query.tracker.DynamoDBQueryTracker;
+import sleeper.query.tracker.QueryState;
+import sleeper.query.tracker.TrackedQuery;
+import sleeper.query.tracker.exception.QueryTrackerException;
 import sleeper.systemtest.drivers.instance.SleeperInstanceContext;
 
+import java.time.Duration;
 import java.util.Map;
 
 import static sleeper.configuration.properties.instance.SystemDefinedInstanceProperty.QUERY_QUEUE_URL;
 
 public class SQSQueryDriver {
+    private static final Logger LOGGER = LoggerFactory.getLogger(SQSQueryDriver.class);
     private final String queueUrl;
     private final AmazonSQS sqsClient;
     private final QuerySerDe querySerDe;
+    private final DynamoDBQueryTracker queryTracker;
+    private final PollWithRetries poll = PollWithRetries.intervalAndPollingTimeout(
+            Duration.ofSeconds(2), Duration.ofMinutes(1));
 
-    public SQSQueryDriver(SleeperInstanceContext instance, AmazonSQS sqsClient) {
+    public SQSQueryDriver(SleeperInstanceContext instance, AmazonSQS sqsClient, AmazonDynamoDB dynamoDBClient) {
         this.sqsClient = sqsClient;
         this.queueUrl = instance.getInstanceProperties().get(QUERY_QUEUE_URL);
         this.querySerDe = new QuerySerDe(Map.of(instance.getTableName(), instance.getTableProperties().getSchema()));
+        this.queryTracker = new DynamoDBQueryTracker(instance.getInstanceProperties(), dynamoDBClient);
     }
 
-    public void run(Query query) {
+    public void run(Query query) throws InterruptedException {
         sqsClient.sendMessage(queueUrl, querySerDe.toJson(query));
+        waitForQuery(query.getQueryId());
+    }
+
+    public void waitForQuery(String queryId) throws InterruptedException {
+        poll.pollUntil("query is finished", () -> {
+            try {
+                TrackedQuery queryStatus = queryTracker.getStatus(queryId);
+                if (queryStatus == null) {
+                    LOGGER.info("Query not found yet, retrying...");
+                    return false;
+                }
+                QueryState state = queryStatus.getLastKnownState();
+                if (QueryState.FAILED == state || QueryState.PARTIALLY_FAILED == state) {
+                    throw new IllegalStateException("Query failed: " + queryStatus);
+                }
+                LOGGER.info("Query found with state: {}", state);
+                return QueryState.COMPLETED == state;
+            } catch (QueryTrackerException e) {
+                throw new RuntimeException(e);
+            }
+        });
     }
 }
