+++ java/system-test/system-test-suite/src/main/java/sleeper/systemtest/suite/dsl/query/SystemTestSQSQuery.java
@@ -23,7 +23,6 @@ import sleeper.systemtest.drivers.query.QueryCreator;
 import sleeper.systemtest.drivers.query.QueryRange;
 import sleeper.systemtest.drivers.query.S3ResultsDriver;
 import sleeper.systemtest.drivers.query.SQSQueryDriver;
-import sleeper.systemtest.drivers.query.WaitForQueryDriver;
 import sleeper.systemtest.suite.fixtures.SystemTestClients;
 
 import java.util.List;
@@ -33,13 +32,11 @@ public class SystemTestSQSQuery {
     private final QueryCreator queryCreator;
     private final SQSQueryDriver sqsQueryDriver;
     private final S3ResultsDriver s3ResultsDriver;
-    private final WaitForQueryDriver waitForQueryDriver;
 
     public SystemTestSQSQuery(SleeperInstanceContext instance, SystemTestClients clients) {
         this.queryCreator = new QueryCreator(instance);
-        this.sqsQueryDriver = new SQSQueryDriver(instance, clients.getSqs());
+        this.sqsQueryDriver = new SQSQueryDriver(instance, clients.getSqs(), clients.getDynamoDB());
         this.s3ResultsDriver = new S3ResultsDriver(instance, clients.getS3());
-        this.waitForQueryDriver = new WaitForQueryDriver(instance, clients.getDynamoDB());
     }
 
     public List<Record> allRecordsInTable() throws InterruptedException {
@@ -52,7 +49,6 @@ public class SystemTestSQSQuery {
 
     private List<Record> runAndReturn(Query query) throws InterruptedException {
         sqsQueryDriver.run(query);
-        waitForQueryDriver.waitForQuery(query.getQueryId());
         return s3ResultsDriver.results(query.getQueryId())
                 .collect(Collectors.toUnmodifiableList());
     }
